+++ src/format/BUILD
@@ -5,6 +5,7 @@ go_library(
     deps = [
         "//src/core",
         "//src/fs",
+        "//src/plz",
         "//src/utils",
         "//third_party/go:buildtools",
         "//third_party/go:logging",
