+++ src/plz/BUILD
@@ -11,7 +11,6 @@ go_library(
         "//src/parse",
         "//src/remote",
         "//src/test",
-        "//src/utils",
         "//third_party/go:logging",
     ],
 )
